[RSV-50] Refactor double to big decimal when using money (#53)
* refactor double to big decimal

* small change

* reservation dto now has big decimal

* half up instead of ceiling

* correction review, ready for merge and squash baby

+++ reservation-api/src/main/java/ca/ulaval/glo4002/reservation/api/reservation/dto/ReservationDto.java
@@ -1,10 +1,11 @@
 package ca.ulaval.glo4002.reservation.api.reservation.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class ReservationDto {
   private String dinnerDate;
-  private double reservationPrice;
+  private BigDecimal reservationPrice;
   private List<CustomerDto> customers;
 
   public String getDinnerDate() {
@@ -15,11 +16,11 @@ public class ReservationDto {
     this.dinnerDate = dinnerDate;
   }
 
-  public double getReservationPrice() {
+  public BigDecimal getReservationPrice() {
     return reservationPrice;
   }
 
-  public void setReservationPrice(double reservationPrice) {
+  public void setReservationPrice(BigDecimal reservationPrice) {
     this.reservationPrice = reservationPrice;
   }
